Added the database schema for multiuser
Added a tables Users and privileges.
Privileges has four fields: user_id,
albums_id, read, write. The id fields
are foreign keys which link to the ids
in the users and albums tables (which
is why albums table is converted to InnoDB).
This structure allows for simple join queries
to get either which users have access to
a specific album, or which albums a specific
user has access to.

I also added a update script, but this has
not been tested yet, since I don't really
understand the update mechanics.